+++ sw.js
@@ -35,6 +35,7 @@ const urlsToCache = [
     './js/modules/clipboardManager.js',
     './js/modules/keyboard.js',
     './js/modules/permissions.js',
+    './js/modules/crashMonitor.js',
     './js/apps/appStore/main.js',
     './js/apps/browser/main.js',
     './js/apps/calculator/main.js',